requirements.txt
python-dotenv
pandas
SQLAlchemy
requests
scikit-learn
matplotlib
psycopg2